split up spotify auth and api
